+++ apps/meteor/client/components/message/header/MessageRoles.stories.tsx
@@ -1,5 +1,5 @@
 import { Box } from '@rocket.chat/fuselage';
-import type { Meta, StoryFn } from '@storybook/react';
+import type { StoryObj, Meta } from '@storybook/react';
 
 import MessageRoles from './MessageRoles';
 
@@ -8,13 +8,15 @@ export default {
 	decorators: [(fn) => <Box>{fn()}</Box>], // TODO: workaround for missing Fuselage default stylesheet
 } satisfies Meta<typeof MessageRoles>;
 
-export const AdministratorExample: StoryFn<typeof MessageRoles> = (args) => <MessageRoles {...args} />;
-AdministratorExample.args = {
-	roles: ['admin', 'user'],
+export const AdministratorExample: StoryObj<typeof MessageRoles> = {
+	args: {
+		roles: ['admin', 'user'],
+	},
 };
 
-export const BotExample: StoryFn<typeof MessageRoles> = (args) => <MessageRoles {...args} />;
-BotExample.args = {
-	roles: ['user'],
-	isBot: true,
+export const BotExample: StoryObj<typeof MessageRoles> = {
+	args: {
+		roles: ['user'],
+		isBot: true,
+	},
 };