+++ packages/ui-voip/src/providers/MediaCallProvider.tsx
@@ -1,7 +1,10 @@
+import { AnchorPortal } from '@rocket.chat/ui-client';
 import type { ReactNode } from 'react';
 
 import MediaCallInstanceProvider from './MediaCallInstanceProvider';
 import MediaCallViewProvider from './MediaCallViewProvider';
+import { MediaCallWidget } from '../views';
+import MediaCallPopout from '../views/MediaCallPopout';
 
 type MediaCallProviderProps = {
 	children: ReactNode;
@@ -10,7 +13,12 @@ type MediaCallProviderProps = {
 const MediaCallProvider = ({ children }: MediaCallProviderProps) => {
 	return (
 		<MediaCallInstanceProvider>
-			<MediaCallViewProvider />
+			<MediaCallViewProvider>
+				<AnchorPortal id='rcx-media-call-widget-portal'>
+					<MediaCallWidget />
+				</AnchorPortal>
+			</MediaCallViewProvider>
+			<MediaCallPopout />
 			{children}
 		</MediaCallInstanceProvider>
 	);